refactor: typescript snyk-test for npm and yarn
